refactor(codemode): split interpreter and stdlib
